refactor: extract helpers
- `isCSSVarProp` from `stylable-utils` to `helpers/css-custom-property`
- `validateAllowedNodesUntil` from `stylable-value-parser` to `helpers/value`
- `globalValue` from `utils` to `helpers/global`
- deprecate `paramMapping` constant export
