+++ packages/core/src/functions.ts
@@ -8,7 +8,7 @@ import { assureRelativeUrlPrefix } from './stylable-assets';
 import type { StylableMeta } from './stylable-meta';
 import type { CSSResolve, JSResolve, StylableResolver } from './stylable-resolver';
 import type { replaceValueHook, StylableTransformer } from './stylable-transformer';
-import { isCSSVarProp } from './stylable-utils';
+import { isCSSVarProp } from './helpers/css-custom-property';
 import { strategies, valueMapping } from './stylable-value-parsers';
 import { getFormatterArgs, getStringValue } from './helpers/value';
 import type { ParsedValue } from './types';
